# app/Models/Conta.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Conta extends Model
{
    use HasFactory;

    protected $table = 'contas';
    protected $fillable = [
        'numero_conta',
        'saldo',
    ];

    protected $casts = [
        'saldo' => 'decimal:2'
    ];
}